Day2/requirements.txt
# Medical AI Agent Web UI Requirements
streamlit>=1.56.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0